requirements.txt
fastapi==0.61.1
uvicorn==0.12.2
jinja2==2.11.2
aiofiles==0.5.0
fastapi==0.61.1
motor==2.2.0
pydantic[email]
python-decouple==3.3
